Fixing SIGSEGV on the ARM

+++ src/Sparrow/Platform/Posix/PosixThreadsMethods.cs
@@ -8,13 +8,13 @@ namespace Sparrow
         private const string LIBC_6 = "libc.so.6";
 
         [DllImport(LIBC_6, CallingConvention = CallingConvention.Cdecl, CharSet = CharSet.Unicode, SetLastError = true)]
-        public static unsafe extern int pthread_setschedparam(ulong thread, int policy, sched_param* param);
+        public static unsafe extern int pthread_setschedparam(IntPtr thread, int policy, sched_param* param);
 
         [DllImport(LIBC_6, CallingConvention = CallingConvention.Cdecl, CharSet = CharSet.Unicode, SetLastError = true)]
-        public static unsafe extern int pthread_getschedparam(ulong thread, int* policy, sched_param* param);
+        public static unsafe extern int pthread_getschedparam(IntPtr thread, int* policy, sched_param* param);
 
         [DllImport(LIBC_6, CallingConvention = CallingConvention.Cdecl, CharSet = CharSet.Unicode, SetLastError = true)]
-        public static extern ulong pthread_self();
+        public static extern IntPtr pthread_self();
     }
 
     public struct sched_param
